backend/requirements.txt
asgiref==3.2.10
certifi==2019.6.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cycler==0.10.0
Cython==0.29.17
Django==3.1.2
djangorestframework==3.11.0
djangorestframework-simplejwt==4.3.0
drf-yasg==1.17.1
httpie==1.0.3
idna==2.8
imageio==2.5.0
inflection==0.4.0
itypes==1.1.0
Jinja2==2.10.1
joblib==0.16.0
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.0.3
nltk==3.4.5
numpy==1.17.1
oauthlib==3.1.0
openapi-codec==1.3.2
packaging==20.3
pandas==0.24.2
Pillow==7.1.1
psycopg2==2.8.3
psycopg2-binary==2.8.3
py4j==0.10.7
Pygments==2.4.2
PyJWT==1.7.1
pyparsing==2.4.2
PySocks==1.7.1
pyspark==2.4.6
PyStemmer==2.0.0.1
python-dateutil==2.8.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
scikit-learn==0.23.1
scipy==1.5.1
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.1
threadpoolctl==2.1.0
tweepy==3.8.0
tzlocal==2.0.0
Unidecode==1.1.1
uritemplate==3.0.0
urllib3==1.25.3
wordcloud==1.5.0
